requirements.txt
flask-wtf
flask-sqlalchemy
flask-migrate
flask-admin